requirements.txt
Django==4.1.10
gunicorn==20.1.0
django-allauth==0.53.1
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
dj-database-url==1.3.0
boto3==1.26.135
s3transfer==0.6.1
botocore==1.29.135
django-storages==1.13.2
psycopg2-binary==2.9.6
Pillow==9.4.0
pytz==2022.7
django-serverless-cron==0.1.3